renamed files and put in new directory
